Igra se pokreče

+++ product-manager/src/main/java/hr/tis/academy/model/dto/GameStateDto.java
@@ -3,14 +3,14 @@ package hr.tis.academy.model.dto;
 import hr.tis.academy.enums.GameState;
 
 public class GameStateDto {
-    public Integer userId;
+    public Long userId;
     public GameState gameState;
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
