+++ product-manager/src/main/java/hr/tis/academy/repository/DuckyUsersRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DuckyUsersRepository extends JpaRepository<DuckyUsers, Integer> {
+public interface DuckyUsersRepository extends JpaRepository<DuckyUsers, Long> {
 }
